Moved to alternating pattern for the front panel input (TrigT1CTP-03-00-76)
	* Moved to alternating pattern for the front panel input
	* TrigT1CTP-03-00-75/6

2015-09-21 Joerg Stelzer
	* Restore some histogramming functionality accidentaly removed
	* TrigT1CTP-03-00-74

2015-09-20 Joerg Stelzer
	* Move histogram booking from start to beginRun
	* TrigT1CTP-03-00-72/73

2015-09-20 Joerg Stelzer
	* Remove zero bias trigger from simulation
	* TrigT1CTP-03-00-71

2015-09-17 Joerg Stelzer
	* Fix CTP simulation
	* TrigT1CTP-03-00-70

2015-08-09 Joerg Stelzer
...
(Long ChangeLog diff - truncated)

+++ Trigger/TrigT1/TrigT1CTP/src/CTPUtil.cxx
@@ -60,7 +60,7 @@ namespace LVL1CTP {
 
   }
 
-  int CTPUtil::getMult( unsigned int word, int startbit, int endbit ) {
+  int CTPUtil::getMult( uint64_t word, unsigned int startbit, unsigned int endbit ) {
 
     return static_cast<int>( ( word >> startbit ) & static_cast<unsigned int>( pow( 2, endbit - startbit + 1 ) - 1 ) );
